+++ src/DTO/Environment/ArchitectureKind.php
@@ -10,6 +10,7 @@ namespace PHPeek\SystemMetrics\DTO\Environment;
 enum ArchitectureKind: string
 {
     case X86_64 = 'x86_64';
+    case X86 = 'x86';
     case Arm64 = 'arm64';
     case Other = 'other';
 }